Added node zlib, socket and http tests
